tests: moved out fixtures
