+++ tests/test_unit/test_cdf_tk/test_client/test_instances.py
@@ -3,44 +3,11 @@ import socket
 
 import pytest
 import responses
-from cognite.client import global_config
-from cognite.client.credentials import Token
 from cognite.client.data_classes.data_modeling import NodeApply, NodeOrEdgeData, ViewId
 from cognite.client.data_classes.data_modeling.cdm.v1 import CogniteTimeSeriesApply
 from cognite.client.exceptions import CogniteAPIError, CogniteConnectionError, CogniteReadTimeout
 
 from cognite_toolkit._cdf_tk.client import ToolkitClient, ToolkitClientConfig
-
-BASE_URL = "http://blabla.cognitedata.com"
-TOKEN_URL = "https://test.com/token"
-
-
-@pytest.fixture
-def toolkit_config():
-    return ToolkitClientConfig(
-        client_name="test-client",
-        project="test-project",
-        base_url=BASE_URL,
-        max_workers=1,
-        timeout=10,
-        credentials=Token("abc"),
-    )
-
-
-@pytest.fixture
-def disable_gzip():
-    old = global_config.disable_gzip
-    global_config.disable_gzip = True
-    yield
-    global_config.disable_gzip = old
-
-
-@pytest.fixture
-def max_retries_2():
-    old = global_config.max_retries
-    global_config.max_retries = 2
-    yield
-    global_config.max_retries = old
 
 
 @pytest.fixture()
